Fix scss variables import in common lib
We don't want big relative paths for scss variables import, and we don't want errors in our IDEs which are not finding the global `@import 'variables';` in our libs (because variables is defined in each app).
Knowing our variables are always the same whatever the app, we decided to reference the absolute path of dashboard's variables in commin lib.