[sc-4636]: Fix how we compute user selection position
We were using a regexp which was finding the first match of the selection in the paragraph, so as soon as we had another occurence of the selection before, the position returned was wrong.
This fix is using properly the Range properties: we compute the selection's position based on the previous mark.