[sc-6131] Refresh configuration form once account changes are saved
When the user set 0 `max_dedicated_processors` for an account that has dedicated processors ENABLED, the rule to take in consideration is that an account with enabled dedicated processors must have at least 1 max dedicated processor. On the server side this value is automatically coerced to 1 to comply with the rule.