[sc-7162]: prevent save & cancel buttons to be enabled by default
We already have the footer’s button being disabled when the form was invalid or pristine, but disabling the form by default and enabling it after getting the data was marking the form as touched, and so enabling the footer’s buttons.
So now, instead of disabling the form we hide it until we get the data to fill it.