[sc-8108] Fix load more results in widget generator page
In the Search Result widget, we were loading more results on scroll but listening only the document. So load more didn't work when the scrollable container was another block of the page.

We added an optional attribute `scrollableContainerSelector` on nuclia-search-results, allowing to specify the query selector of the scrollable block if any.