[sc-8108] and [sc 8198]: widget generator fixes (#1231)
* [sc-8198] Move private KB warning in code snippet overlay

Display a confirmation dialog when copying a private KB snippet

* [sc-8108] Fix load more results in widget generator page

In the Search Result widget, we were loading more results on scroll but listening only the document. So load more didn't work when the scrollable container was another block of the page.

We added an optional attribute `scrollableContainerSelector` on nuclia-search-results, allowing to specify the query selector of the scrollable block if any.