[sc-8106] Render the resource title as link in resource list
In resource list, we created a new `title-cell` component because all the resource tables (processed, pending and error) are sharing the same code for the title cell.

Now the resource title is a real link (but we kept the same style, except it is underlined on hover), so we can easily open it on a new tab. Clicking anywhere else on the title cell is still navigating to the resource on the same tab as it used to.