Upgrade to angular 17.3.2 and nx 18.2.1 (#1450)
Nx migration step failed because of this bug: https://github.com/nrwl/nx/issues/20681. But I feel like it wasn't a big deal for this minor version change. The most important thing was the deps to be upgrading to fix dependabot issues. That being said, we might have problem on the next major upgrade because of this bug, so we should follow it up.