[sc-10815]: Fix feature flag service (#1649)
To fix https://app.shortcut.com/flaps/story/10124/feature-flags-not-working-in-onboarding-page, we added a `startWith('')` on `accountMd5` of `FeatureFlagService` to make it work also on onboarding page.

The problem with this first fix, is that `startWith` was in a `combineLatest`, so `accountMd5` was always sending `''` first whenever `applicationRemoteFeatures` was called. This first call with `accountMd5 = ''` introduced the bug https://app.shortcut.com/flaps/story/10815/creating-a-pinecone-kb-from-the-ui-doesn-t-work-in-prod where the pinecone form was displayed but then the form value wasn't sent to the KB creation payload. This is because when submitting the form, we are checking the feature flag with a `take(1)` and the first call was always returning `false` because of this `startWith('')`.